fix: resolve DeepSource Python findings
