+++ tests/integration/test_migrate_database_operator.py
@@ -51,7 +51,7 @@ def test_migration_rejects_store_with_only_disabled_user(tmp_path) -> None:
     with pytest.raises(RuntimeError, match="credential provisioning incomplete"):
         migrate_configured_databases(settings)
 
-    with closing(sqlite3.connect(auth_path)) as connection, connection:
+    with closing(sqlite3.connect(auth_path)) as connection:
         row = connection.execute(
             "SELECT disabled, hashed_password FROM user_credentials WHERE username = ?",
             ("disabled-admin",),
